Properly populate subscription ID column when sending MMS.
Populate the DB column, and feed the correct value into it for the !MSIM
case. Also clean up the code a little to make the subscription ID value
flow a little more readable.

+++ src/com/android/mms/transaction/SmsReceiverService.java
@@ -295,13 +295,12 @@ public class SmsReceiverService extends Service {
     private void handleServiceStateChanged(Intent intent) {
         // If service just returned, start sending out the queued messages
         ServiceState serviceState = ServiceState.newFromBundle(intent.getExtras());
-        long subId = intent.getLongExtra(PhoneConstants.SUBSCRIPTION_KEY, 0);
-        long prefSubId = SubscriptionManager.getDefaultSmsSubId();
+        int subId = intent.getIntExtra(PhoneConstants.SUBSCRIPTION_KEY, 0);
+        int prefSubId = SubscriptionManager.getDefaultSmsSubId();
         // if service state is IN_SERVICE & current subscription is same as
         // preferred SMS subscription.i.e.as set under SIM Settings, then
         // sendFirstQueuedMessage.
-        if (serviceState.getState() == ServiceState.STATE_IN_SERVICE &&
-            subId == prefSubId) {
+        if (serviceState.getState() == ServiceState.STATE_IN_SERVICE && subId == prefSubId) {
             sendFirstQueuedMessage();
         }
     }
